requirements.txt
census==0.8.22
certifi==2024.8.30
charset-normalizer==3.3.2
geographiclib==2.0
geopandas==1.0.1
geopy==2.4.1
idna==3.8
jellyfish==1.1.0
numpy==2.1.1
packaging==24.1
pandas==2.2.2
pyogrio==0.9.0
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
setuptools==74.1.2
shapely==2.0.6
six==1.16.0
tzdata==2024.1
urllib3==2.2.2
us==3.2.0
